refactor: update user api implementation with decorators
